extracted ViewerColumn props
