+++ src/App.tsx
@@ -15,7 +15,13 @@ import { useSessionLoadModal } from './hooks/useSessionLoadModal';
 import { useUpdateChainInfo } from './hooks/useUpdateChainInfo';
 import { useUpdateResidueInfo } from './hooks/useUpdateResidueInfo';
 import { useUpdateColors } from './hooks/useUpdateColors';
-import ViewerColumn from './components/ViewerColumn';
+import ViewerColumn , {
+    getLoadDataRowProps,
+    getMoleculeUIAlignedToProps,
+    getMoleculeUIAlignedProps,
+    getRealignedMoleculeListProps,
+    getMolstarContainerProps
+} from './components/ViewerColumn';
 import TwoColumnsContainer from './components/TwoColumnsContainer';
 import AppHeader from './components/AppHeader';
 import { AlignedTo, Aligned, ReAligned } from './types/molecule';
@@ -740,448 +746,7 @@ const App: React.FC = () => {
      * @returns The properties for the load data row of the viewer, including handlers for file input,
      * representation selection, color addition, subunit and chain selection, and fog/camera controls.
      */
-    function getLoadDataRowProps({
-        viewer,
-        otherViewer,
-        molstar,
-        otherMolstar,
-        realignedStructRefs,
-        otherRealignedStructRefs,
-        isMoleculeAlignedLoaded,
-        isMoleculeAlignedToLoaded,
-        viewerReady,
-        otherViewerReady,
-        representationType,
-        setRepresentationType,
-        colorsFile,
-        isMoleculeColoursLoaded,
-        structureRef,
-        otherStructureRef,
-        selectedSubunit,
-        setSelectedSubunit,
-        subunitToChainIds,
-        chainInfo,
-        selectedChainId,
-        setSelectedChainId,
-        residueInfo,
-        selectedResidueId,
-        setSelectedResidueId,
-        fog,
-        setFog,
-        camera,
-        setCamera,
-        updateFog,
-        handleFileChange,
-        Aligned,
-        allowedRepresentationTypes,
-        syncEnabled,
-        realignedRepRefs,
-        setRealignedRepRefs,
-        setRealignedStructRefs
-    }: LoadDataRowPropsInput) {
-        return {
-            viewerTitle: viewer.moleculeAligned ? Aligned + `: ${viewer.moleculeAligned.name || viewer.moleculeAligned.filename}` : "",
-            isLoaded: isMoleculeAlignedLoaded,
-            onFileInputClick: viewer.handleFileInputButtonClick,
-            fileInputRef: viewer.fileInputRef,
-            onFileChange: (e: any) => handleFileChange(e, Aligned),
-            fileInputDisabled: !isMoleculeAlignedToLoaded || !viewerReady || !otherViewerReady,
-            fileInputLabel: `Load ${Aligned}`,
-            representationType,
-            onRepresentationTypeChange: setRepresentationType,
-            representationTypeDisabled: !isMoleculeAlignedLoaded,
-            representationTypeSelector: (
-                <RepresentationSelectButton
-                    label="Select Representation"
-                    options={[...allowedRepresentationTypes]}
-                    selected={representationType}
-                    onSelect={option => setRepresentationType(option as AllowedRepresentationType)}
-                    disabled={!isMoleculeAlignedLoaded}
-                />
-            ),
-            onAddColorsClick: colorsFile.handleButtonClick,
-            addColorsDisabled: !isMoleculeAlignedLoaded,
-            onAddRepresentationClick: () => {
-                let colorTheme;
-                if (isMoleculeColoursLoaded) {
-                    colorTheme = { name: Aligned + '-custom-chain-colors', params: {} };
-                } else {
-                    colorTheme = { name: 'default', params: {} };
-                }
-                const repId = (typeof crypto !== 'undefined' && crypto.randomUUID ? crypto.randomUUID() : Math.random().toString(36).slice(2));
-                if (viewer.moleculeAligned && structureRef) {
-                    molstar.addRepresentation(
-                        Aligned,
-                        structureRef,
-                        representationType,
-                        colorTheme,
-                        repId
-                    );
-                }
-                if (otherViewer.moleculeAligned && otherStructureRef) {
-                    otherMolstar.addRepresentation(
-                        Aligned,
-                        otherStructureRef,
-                        representationType,
-                        colorTheme,
-                        repId
-                    );
-                }
-                Object.entries(realignedStructRefs).forEach(([id, structRef]) => {
-                    if (structRef) {
-                        molstar.addRepresentation(
-                            id,
-                            structRef,
-                            representationType,
-                            colorTheme,
-                            repId
-                        );
-                    }
-                });
-                Object.entries(otherRealignedStructRefs).forEach(([id, structRef]) => {
-                    if (structRef) {
-                        otherMolstar.addRepresentation(
-                            id,
-                            structRef,
-                            representationType,
-                            colorTheme,
-                            repId
-                        );
-                    }
-                });
-            },
-            addRepresentationDisabled: !isMoleculeAlignedLoaded || !structureRef,
-            colorsInputRef: colorsFile.inputRef,
-            onColorsFileChange: colorsFile.handleFileChange,
-            selectedSubunit,
-            onSelectSubunit: setSelectedSubunit,
-            subunitSelectDisabled: !isMoleculeAlignedLoaded,
-            chainInfo,
-            selectedChainId,
-            onSelectChainId: setSelectedChainId,
-            chainSelectDisabled: !isMoleculeAlignedToLoaded,
-            residueInfo,
-            selectedResidueId,
-            onSelectResidueId: setSelectedResidueId,
-            residueSelectDisabled: !isMoleculeAlignedLoaded,
-            fogEnabled: fog.enabled,
-            fogNear: fog.near,
-            fogFar: fog.far,
-            onFogEnabledChange: (val: boolean) => {
-                setFog.setEnabled(val);
-                updateFog(viewer.ref.current, otherViewer.ref.current, val, fog.near, fog.far, camera.near, camera.far);
-            },
-            onFogNearChange: (val: number) => {
-                setFog.setNear(val);
-                updateFog(viewer.ref.current, otherViewer.ref.current, fog.enabled, val, fog.far, camera.near, camera.far);
-            },
-            onFogFarChange: (val: number) => {
-                setFog.setFar(val);
-                updateFog(viewer.ref.current, otherViewer.ref.current, fog.enabled, fog.near, val, camera.near, camera.far);
-            },
-            cameraNear: camera.near,
-            cameraFar: camera.far,
-            onCameraNearChange: (val: number) => {
-                setCamera.setNear(val);
-                updateFog(viewer.ref.current, otherViewer.ref.current, fog.enabled, fog.near, fog.far, val, camera.far);
-            },
-            onCameraFarChange: (val: number) => {
-                setCamera.setFar(val);
-                updateFog(viewer.ref.current, otherViewer.ref.current, fog.enabled, fog.near, fog.far, camera.near, val);
-            },
-            subunitToChainIds
-        };
-    }
 
-
-    /**
-     * Returns the props for the aligned molecule list component for a viewer (generalized).
-     * @param params Object containing all necessary props for the viewer and its counterpart.
-     * @returns Props for the aligned molecule list component for the viewer.
-     */
-    function getMoleculeUIAlignedToProps({
-        molstar,
-        otherMolstar,
-        viewer,
-        isVisible,
-        onToggleVisibility,
-        chainZoomLabel,
-        onChainZoom,
-        chainZoomDisabled,
-        residueZoomLabel,
-        onResidueZoom,
-        residueZoomDisabled,
-        isLoaded,
-        forceUpdate,
-        representationRefs,
-        syncEnabled,
-        deleteRepresentation,
-        repIdMap,
-        AlignedTo
-    }: {
-        molstar: any,
-        otherMolstar: any,
-        viewer: any,
-        isVisible: boolean,
-        onToggleVisibility: () => void,
-        chainZoomLabel: string,
-        onChainZoom: () => void,
-        chainZoomDisabled: boolean,
-        residueZoomLabel: string,
-        onResidueZoom: () => void,
-        residueZoomDisabled: boolean,
-        isLoaded: boolean,
-        forceUpdate: () => void,
-        representationRefs: any[],
-        syncEnabled: boolean,
-        deleteRepresentation: any,
-        repIdMap: any,
-        AlignedTo: string
-    }) {
-        return {
-            key: representationRefs?.join('-') || viewer.viewerKey + `-` + AlignedTo,
-            label: viewer.moleculeAlignedTo?.label ?? AlignedTo,
-            plugin: viewer.ref.current,
-            isVisible,
-            onToggleVisibility,
-            chainZoomLabel,
-            onChainZoom,
-            chainZoomDisabled,
-            residueZoomLabel,
-            onResidueZoom,
-            residueZoomDisabled,
-            isLoaded,
-            forceUpdate,
-            representationRefs: representationRefs || [],
-            onDeleteRepresentation: (ref: string) => {
-                const repId = Object.entries(repIdMap[AlignedTo]).find(([id, r]) => r === ref)?.[0];
-                if (syncEnabled && repId) {
-                    Promise.all([
-                        deleteRepresentation(molstar.repIdMap[AlignedTo][repId], AlignedTo, molstar, false),
-                        deleteRepresentation(otherMolstar.repIdMap[AlignedTo][repId], AlignedTo, otherMolstar, false)
-                    ]).then(forceUpdate);
-                } else if (repId) {
-                    deleteRepresentation(molstar.repIdMap[AlignedTo][repId], AlignedTo, molstar);
-                } else {
-                    if (syncEnabled) {
-                        Promise.all([
-                            deleteRepresentation(ref, AlignedTo, molstar, false),
-                            deleteRepresentation(ref, AlignedTo, otherMolstar, false)
-                        ]).then(forceUpdate);
-                    } else {
-                        deleteRepresentation(ref, AlignedTo, molstar);
-                    }
-                }
-            },
-            onToggleRepVisibility: (ref: string) => {
-                [molstar, otherMolstar].forEach(molstarInstance => {
-                    const plugin = molstarInstance.pluginRef.current;
-                    if (!plugin) return;
-                    const cell = plugin.state?.data?.cells?.get(ref);
-                    if (cell) {
-                        import('molstar/lib/mol-plugin/commands').then(({ PluginCommands }) => {
-                            PluginCommands.State.ToggleVisibility.apply(plugin, [plugin, { state: plugin.state.data, ref }]);
-                            plugin.canvas3d?.requestDraw?.();
-                            forceUpdate();
-                        });
-                    }
-                });
-            },
-        };
-    }
-
-    /**
-     * Returns the props for the aligned molecule list component for a viewer.
-     * @param params Object containing all necessary props for the viewer and its counterpart.
-     * @returns Props for the aligned molecule list component for the viewer.
-     */
-    function getMoleculeUIAlignedProps({
-        molstar,
-        otherMolstar,
-        viewer,
-        isVisible,
-        onToggleVisibility,
-        chainZoomLabel,
-        onChainZoom,
-        chainZoomDisabled,
-        residueZoomLabel,
-        onResidueZoom,
-        residueZoomDisabled,
-        isLoaded,
-        forceUpdate,
-        representationRefs,
-        syncEnabled,
-        deleteRepresentation,
-        repIdMap,
-        Aligned,
-        chainInfoAligned,
-        selectedChainIdAligned,
-        residueInfoAligned,
-        selectedResidueIdAligned
-    }: {
-        molstar: any,
-        otherMolstar: any,
-        viewer: any,
-        isVisible: boolean,
-        onToggleVisibility: () => void,
-        chainZoomLabel: string,
-        onChainZoom: () => void,
-        chainZoomDisabled: boolean,
-        residueZoomLabel: string,
-        onResidueZoom: () => void,
-        residueZoomDisabled: boolean,
-        isLoaded: boolean,
-        forceUpdate: () => void,
-        representationRefs: any[],
-        syncEnabled: boolean,
-        deleteRepresentation: any,
-        repIdMap: any,
-        Aligned: string,
-        chainInfoAligned: any,
-        selectedChainIdAligned: any,
-        residueInfoAligned: any,
-        selectedResidueIdAligned: any
-    }) {
-        return {
-            key: representationRefs?.join('-') || viewer.viewerKey + `-` + Aligned,
-            label: viewer.moleculeAligned?.label ?? Aligned,
-            plugin: viewer.ref.current,
-            isVisible,
-            onToggleVisibility,
-            chainZoomLabel,
-            onChainZoom,
-            chainZoomDisabled,
-            residueZoomLabel,
-            onResidueZoom,
-            residueZoomDisabled,
-            isLoaded,
-            forceUpdate,
-            representationRefs: representationRefs || [],
-            onDeleteRepresentation: (ref: string) => {
-                const repId = Object.entries(repIdMap[Aligned]).find(([id, r]) => r === ref)?.[0];
-                if (syncEnabled && repId) {
-                    Promise.all([
-                        deleteRepresentation(molstar.repIdMap[Aligned][repId], Aligned, molstar, false),
-                        deleteRepresentation(otherMolstar.repIdMap[Aligned][repId], Aligned, otherMolstar, false)
-                    ]).then(forceUpdate);
-                } else if (repId) {
-                    deleteRepresentation(molstar.repIdMap[Aligned][repId], Aligned, molstar);
-                } else {
-                    if (syncEnabled) {
-                        Promise.all([
-                            deleteRepresentation(ref, Aligned, molstar, false),
-                            deleteRepresentation(ref, Aligned, otherMolstar, false)
-                        ]).then(forceUpdate);
-                    } else {
-                        deleteRepresentation(ref, Aligned, molstar);
-                    }
-                }
-            },
-            onToggleRepVisibility: (ref: string) => {
-                [molstar, otherMolstar].forEach(molstarInstance => {
-                    const plugin = molstarInstance.pluginRef.current;
-                    if (!plugin) return;
-                    const cell = plugin.state?.data?.cells?.get(ref);
-                    if (cell) {
-                        import('molstar/lib/mol-plugin/commands').then(({ PluginCommands }) => {
-                            PluginCommands.State.ToggleVisibility.apply(plugin, [plugin, { state: plugin.state.data, ref }]);
-                            plugin.canvas3d?.requestDraw?.();
-                            forceUpdate();
-                        });
-                    }
-                });
-            },
-        };
-    }
-
-    /**
-     * Returns the props for the realigned molecule list component for a viewer.
-     * @param params Object containing all necessary props for the viewer and its counterpart.
-     * @returns Props for the realigned molecule list component for the viewer.
-     */
-    function getRealignedMoleculeListProps({
-        molecules,
-        molstar,
-        chainInfo,
-        residueInfo,
-        selectedResidueId,
-        realignedStructRefs,
-        setRealignedMolecules,
-        setRealignedRepRefs,
-        setRealignedStructRefs,
-        forceUpdate,
-        viewerKey,
-        otherMolstar,
-        otherRealignedStructRefs,
-        setOtherRealignedMolecules,
-        setOtherRealignedRepRefs,
-        setOtherRealignedStructRefs
-    }: {
-        molecules: any,
-        molstar: any,
-        chainInfo: any,
-        residueInfo: any,
-        selectedResidueId: any,
-        realignedStructRefs: any,
-        setRealignedMolecules: any,
-        setRealignedRepRefs: any,
-        setRealignedStructRefs: any,
-        forceUpdate: () => void,
-        viewerKey: string,
-        otherMolstar: any,
-        otherRealignedStructRefs: any,
-        setOtherRealignedMolecules: any,
-        setOtherRealignedRepRefs: any,
-        setOtherRealignedStructRefs: any
-    }) {
-        return {
-            molecules,
-            molstar,
-            chainInfo,
-            residueInfo,
-            selectedResidueId,
-            realignedStructRefs,
-            setRealignedMolecules,
-            setRealignedRepRefs,
-            setRealignedStructRefs,
-            forceUpdate,
-            viewerKey,
-            otherMolstar,
-            otherRealignedStructRefs,
-            setOtherRealignedMolecules,
-            setOtherRealignedRepRefs,
-            setOtherRealignedStructRefs,
-        };
-    }
-
-    /**
-     * Returns the props for the Molstar container component for a viewer.
-     *
-     * @param viewer The viewer state object (viewerA or viewerB).
-     * @param pluginRef The plugin ref for the viewer.
-     * @param setViewerWrapper The setViewer wrapper for the viewer.
-     * @param setViewerReady The setViewerReady setter for the viewer.
-     * @returns Props for the Molstar container component for the viewer.
-     */
-    function getMolstarContainerProps({
-        viewer,
-        pluginRef,
-        setViewerWrapper,
-        setViewerReady
-    }: {
-        viewer: typeof viewerA,
-        pluginRef: typeof pluginRefA,
-        setViewerWrapper: (viewer: any) => void,
-        setViewerReady: (ready: boolean) => void
-    }) {
-        return {
-            ref: pluginRef,
-            viewerKey: viewer.viewerKey,
-            setViewer: setViewerWrapper,
-            onMouseDown: () => setActiveViewer(viewer.viewerKey),
-            onReady: () => setViewerReady(true),
-        };
-    }
 
     // Return the main app component.
     return (
